requirements.txt
aiogram==2.13
numpy==1.16.4
opencv-python==4.5.2.54
Pillow==8.2.0
tensorflow==1.14.0
